FullTechTest
Tech test done as part of an application to SABS Travel Technologies. Test involved parsing JSON and displaying it as a list. Pretty sure I interpreted the JSON data wrong because the data I was left with seems weird but it still serves as a solid demonstration of data handling skills